refactor: HTTP method 별도 파일로 분리 및 호스트 변수 refactor
