+++ src/resources/Masking.ts
@@ -1,4 +1,4 @@
-import HTTPMethod from '../constants';
+import HTTPMethod from '../HttpMethods';
 import { DataResponse } from '../interfaces';
 import Base from './_Base';
